Refactoring of SearchPage
